pyproject.toml
[project]
name = "smart-link-service"
version = "1.0.0"
description = "SmartLink Agent Management Platform - Backend Service"
authors = [{name = "SmartLink Team", email = "team@smartlink.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    
    # Redis
    "redis[asyncio]>=5.0.1",
    
    # LLM Integration
    "litellm>=1.20.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    
    # MCP
    "mcp>=0.9.0",
    
    # LangChain (基础能力)
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    
    # Validation & Serialization
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    
    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]